Extract function, use consistent types

+++ src/gui/MenuWidgets.h
@@ -492,7 +492,7 @@ public:
 class CMenuCheckButton : public CMenuElement {
 	
 public:
-	CMenuCheckButton(int id, Vec2f pos, int size, TextureContainer * tex1,
+	CMenuCheckButton(int id, Vec2i pos, int size, TextureContainer * tex1,
 					 TextureContainer * tex2, CMenuElementText * label = NULL); 
 	virtual ~CMenuCheckButton();
